requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
